Add files via upload
This study was conducted to determine the pattern of the relationship between the independent variables that affect the Poverty Depth Index in 35 Regencies/Cities in Central Java Province in 2017-2020. The independent variables in this study include the Open Unemployment Rate (TPT), Average Years of Schooling (RLS), Expenditures per Capita (PPK), and District/City Minimum Wages (UMK). The analytical method used in this study is panel data regression analysis.

The analysis in this report uses secondary data obtained from the publications of the Central Statistics Agency (BPS) of Central Java Province. The data taken is annual data for the 2017-2020 period from 35 regencies/cities in Central Java Province. The data in the analysis in this report is panel data which is a combination of cross-sectional data and time series data. The data consists of five variables consisting of one dependent variable and four independent variables. The dependent variable used is the Poverty Depth Index. While the independent variables used are the Open Unemployment Rate (TPT), Average Years of Schooling (RLS), Expenditures per Capita (PPK), and District/City Minimum Wages (UMK).